Move providers directory into src
