+++ frontend/src/pages/_app.tsx
@@ -3,7 +3,7 @@ import type { AppProps } from "next/app";
 import Layout from "../components/common/layout";
 import { Noto_Sans } from "next/font/google";
 import AuthContextProvider from "@/contexts/AuthContext";
-import { MatchmakingProvider } from "../../providers/MatchmakingProvider";
+import { MatchmakingProvider } from "../providers/MatchmakingProvider";
 import AuthChecker from "@/components/common/auth-checker";
 
 const notoSans = Noto_Sans({
